Exclude the backtrace in a failure::Error from the error typename
A failure::Error (and by extension, a failure::Context wrapping a
failure::Error and passed around as a failure::Fail) will print its
backtrace in the Debug output. To get the typename from a Fail, we
need to exclude the backtrace by only using the first line of the
Fail's Debug output.