feat: add support for openssl-less builds
This issue is related to #88 and would prevent issues like it when
openssl 1.2 comes out.

Currently, sentry-rust depends on reqwest with default feature flags,
which pulls in `native-tls` and `openssl`.

This causes multiple issues:

1. On systems without openssl header packages installed, openssl-sys
   won't build.
2. If distributions like arch linux update to newer versions of openssl,
   issues like #88 will arise
3. Systems that don't support openssl are not supported (although
   targeting embedded systems with non-unix OSes is probably not a goal
   in this project)

This commit adds support for building reqwest with rustls support, using
the new feature flag `rustls`.

This fixes #136.