Bump findshlibs to 0.7.0 (#211)
findshlibs 0.5.0 has a serious defect in `SharedLibrary::id`, where it
uses file sizes and offsets when examining the PT_NOTE sections of an
ELF binary instead of the memory size/offset fields. This issue was
corrected by gimli-rs/findshlibs@046a431, but was released along with a
couple of semver breaking changes so it needs a bump in Cargo.toml to be
picked up.