fix: Explicitly drop SessionFlusher
It seems that doing a thread::join from within the Drop impl of a thread
local THREAD_HUB will deadlock for some reason.
Explicitly joining the background flusher as part of the ClientGuard
Drop does work, and is also more correct, as otherwise the PROCESS_HUB
might keep the worker alive forever.