docs: Expose feature-gated APIs with doc(cfg)
APIs hidden behind features can now be viewed on docs.rs!
All feature-gated modules and structs should be annotated
with a note that mentions the specific feature flag needed
to use the corresponding API.

This primarily fixes a linter warning regarding a link to a
nonexistent integration which was gated behind a feature.
There are other small fixes as well, mostly just hooking up
links that weren't assigned destinations.

While the doc feature is considered unstable, we're taking
advantage of the fact that docs are built with nightly by
default.