refactor: Remove tokio dependency for blocking ureq and curl transports (#422)
These transport backends don't need the overhead of a tokio runtime and
all the dependency crates that come with it, when they are already
blocking on their own.  A simple Rust thread and mpsc channel is enough.
The transport thread implementation is duplicated between a tokio and
non-tokio module that are used by reqwest/surf, and curl/ureq
respectively.

See also
https://github.com/getsentry/sentry-rust/pull/419#discussion_r794287640
where this was briefly discussed.