ref: Change the internal representation of project IDs from u64s to strings (#452)
Project IDs aren't guaranteed to be numbers forever, and shouldn't
be treated as such any more, to be forward-compatible with 
whatever new identifier format that does get used in the future for
project IDs. They're now treated as opaque strings internally, and 
are subjected to minimal validation due to their opacity.