fix: Avoid compile error with profiling frame ip
This tries to solve a weird compile error that does not happen in this
repo directly but rather when used from symbolicator.

The reason this fails in one repo but not the other is not quite obvious though.

`ip()` as a method is available via an internal `pprof::backtrace::Frame`
trait that is not public so we can't import it directly.
However, the underlying type when using the `backtrace` backend also
has a *method* `ip()`.
The underlying impl when using the `frame-pointer` feature instead has
a field called `ip`.