Re-disable scheduled jobs on forks
I disabled scheduled jobs on forks in #435 to prevent [unnecessary spam
from steps that are not supposed to succeed anyway], but @Swatinem
removed the check in #528 for an undocumented reason.  Add it back to
not bother forkers with day-to-day issues, while still allowing them to
test/see CI results locally on push (to one of a limited set of
preconfigured branches).

[unnecessary spam from steps that are not supposed to succeed anyway]: https://github.com/MarijnS95/sentry-rust/actions/runs/3926205216