tower: Use `tx.set_request()` to set HTTP request metadata
Without this the transaction does not have any HTTP request metadata attached, because apparently the event processor is not running for performance traces.

Note that I had to reorder the contents of the `configure_scope()` callback to be able to access `sentry_req` before it disappears into the `add_event_processor()` callback.