Use `rand` instead of `getrandom` for IDs
The `getrandom` crate uses syscalls to get *true* randomness from the OS.
We do not need that high quality of randomness, we do not even need cryptographically secure randomness.

This switches to the `rand` crate which uses a reseeding PRNG which is still cryptographically secure though.

This is much faster, and we can switch to an even simpler PRNG if it is not fast enough still.