fix(ClientInitGuard): ensure drop can only be called once
In some cases, we might want / need to clone the ClientInitGuard, or maybe even
send it in many threads. The only way to get it to work today is to call
`std::mem::forget` on it, which is not ideal, as it will never be dropped.
This is because the first time the ClientInitGuard will be dropped, the session
will be ended, and the client closed.

This can be fixed by checking the # of references to the guard in the drop. By
doing so, we can ensure that we close everything only when the last reference
to it is dropped. This makes cases where we want to clone it around possible,
without using `std::mem::forget`, and thus allowing us to keep the drop
behaviour, when needed.