docs: match description of `debug` option with behavior since #820
PR #820 removed the `debug-logs` crate feature, but failed to update the
documentation for the related `debug` field in the client options
struct, which falsely asserts that the `log` feature changes the
behavior of the `sentry_debug!` macro.

This change fixes that discrepancy by removing any reference to features
in the documentation for that field.